+++ frontend/src/main/java/org/abs_models/backend/java/lib/runtime/ABSBuiltInFunctions.java
@@ -46,7 +46,7 @@ public class ABSBuiltInFunctions {
     }
 
     public static ABSRational currentms() {
-        return ABSRational.fromAprational(new Aprational(new Apint(System.currentTimeMillis()), new Apint(1000)));
+        return ABSRational.fromAprational(ABSRuntime.getRuntime().getClock());
     }
 
     public static ABSInteger ms_since_model_start() {