+++ frontend/src/main/java/org/abs_models/backend/java/lib/runtime/ABSGuard.java
@@ -7,16 +7,56 @@ package org.abs_models.backend.java.lib.runtime;
 import org.abs_models.backend.java.observing.FutView;
 import org.abs_models.backend.java.observing.GuardView;
 
+import org.apfloat.Aprational;
+
+/**
+ * The base class of all guards.
+ */
 public abstract class ABSGuard {
+
+    /**
+     * Return whether the guard is true; i.e., whether the process waiting on
+     * the guard can be scheduled.
+     * <p>
+     * NOTE: It is important to know whether to call this method or the {@link
+     * #await} method.  The await method informs the cog about the task's
+     * status, so should be called when the result will lead to a scheduling
+     * decision.  The {@code isTrue} method should be called when we need to
+     * display the guard's status in some way, e.g. for pretty-printing or
+     * debugging purposes.
+     * <p>
+     * To obey this protocol is most important for guards of type {@link
+     * ABSExpGuard}, since that guard changes its value frequently, and the
+     * runtime must be informed about its effective status at all times.
+     *
+     * @return the guard's status.
+     */
     public abstract boolean isTrue();
 
     /**
-     * 
-     * @return whether this guard could become false again in the future
+     * Check if the guard is monotonic, i.e., if its state can oscillate
+     * between false and true or not.
+     *
+     * @return true if guard is monotonic, false if not.
      */
-    public boolean await() {
-        return !staysTrue();
-    }
+    public abstract boolean staysTrue();
+
+    /**
+     * Wait and/or check if the guard is true.
+     * <p>
+     * This method is allowed to suspend the thread, but should arrange to be
+     * woken up, e.g., by the future becoming available.  This method will be
+     * called multiple times, hence should not do too much work once the guard
+     * is true.
+     * <p>
+     * All implementations of this method must inform the cog about their
+     * status via the methods {@link COG#notifyAwait} and {@link
+     * COG#notifyWakeup()}.
+     *
+     * @param cog the cog of the task that might be suspended.
+     * @return the status of the guard (true if runnable, false if not).
+     */
+    public abstract boolean await(COG cog);
 
     private GuardView view;
 
@@ -24,10 +64,6 @@ public abstract class ABSGuard {
         if (view == null)
             view = new View();
         return view;
-    }
-
-    public boolean staysTrue() {
-        return true;
     }
 
     private class View implements GuardView {
@@ -44,6 +80,10 @@ public abstract class ABSGuard {
             return ABSGuard.this instanceof ABSFutureGuard;
         }
 
+        public boolean isDurationGuard() {
+            return ABSGuard.this instanceof ABSDurationGuard;
+        }
+
         public boolean isAndGuard() {
             return ABSGuard.this instanceof ABSAndGuard;
         }
@@ -58,6 +98,14 @@ public abstract class ABSGuard {
 
         public FutView getFuture() {
             return ((ABSFutureGuard) ABSGuard.this).fut.getView();
+        }
+
+        public Aprational getMinTime() {
+            return ((ABSDurationGuard) ABSGuard.this).getMinTime();
+        }
+
+        public Aprational getMaxTime() {
+            return ((ABSDurationGuard) ABSGuard.this).getMaxTime();
         }
 
         @Override