+++ frontend/src/main/java/org/abs_models/backend/java/lib/types/ABSRational.java
@@ -145,6 +145,10 @@ public class ABSRational extends ABSBuiltInDataType {
         return value.doubleValue();
     }
 
+    public Aprational toAprational() {
+        return value;
+    }
+
     @Override
     public String toString() {
         return value.toString();